pyproject.toml
[project]
name = "sbbd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-celery-beat>=2.1.0",
    "django-celery-results>=2.6.0",
    "django-debug-toolbar>=6.2.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.1.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pip>=26.0.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.1",
]

[tool.ruff]
# 한 줄의 최대 길이 (보통 88 또는 100 사용)
line-length = 120

# 검사할 규칙 설정 (E, F는 기본, I는 import 정렬)
lint.select = ["E", "F", "I", "N", "UP"]
lint.ignore = []

# 절대 삭제하면 안 되는 자동 수정 기능 설정
lint.fixable = ["ALL"]

[tool.ruff.lint.isort]
# 장고 프로젝트에서 import 순서를 예쁘게 맞춰줍니다.
known-first-party = ["config", "apps"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 120

[tool.ruff.lint.per-file-ignores]
"config/settings/*.py" = ["F403", "F405"]
